+++ src/buffer/mesh_triangle.hpp
@@ -31,7 +31,7 @@ namespace VTX
 			void setPositions( const std::vector<Vec3f> & );
 			void setNormals( const std::vector<Vec3f> & );
 			void setColors( const std::vector<Color::Rgb> & );
-			void setVisibilities( const std::vector<ushort> & );
+			void setVisibilities( const std::vector<uint> & );
 			void setIndices( const std::vector<uint> & );
 
 		  protected:
