+++ src/shader/cylinder.frag
@@ -10,7 +10,7 @@ in GsOut
 	smooth vec3			viewImpostorPosition; // Impostor position in view space.
 	flat vec3			viewVertices[ 2 ];	  // Cylinder vertices position in view space.
 	flat vec3			colors[ 2 ];
-	flat unsigned short vertexSelected[ 2 ];
+	flat uint vertexSelected[ 2 ];
 }
 gsIn;
