+++ src/shader/cylinder.vert
@@ -9,8 +9,8 @@
 layout( location = ATOM_POSITION ) in vec3 aVertexPosition;
 layout( location = ATOM_COLOR ) in vec3 aVertexColor;
 layout( location = ATOM_RADIUS ) in float aVertexRad;
-layout( location = ATOM_VISIBILITY ) in unsigned short aVertexVis;
-layout( location = ATOM_SELECTION ) in unsigned short aVertexSel;
+layout( location = ATOM_VISIBILITY ) in uint aVertexVis;
+layout( location = ATOM_SELECTION ) in uint aVertexSel;
 
 uniform mat4 u_MVMatrix;
 uniform mat4 u_projMatrix;
@@ -18,8 +18,8 @@ uniform mat4 u_projMatrix;
 out VsOut
 {
 	flat vec3			vertexColor;
-	flat unsigned short vertexVisible;
-	flat unsigned short vertexSelected;
+	flat uint vertexVisible;
+	flat uint vertexSelected;
 }
 vsOut;
