+++ src/shader/ribbon_patch.vert
@@ -2,17 +2,17 @@
 
 layout( location = 0 ) in vec4 vao_controlPointPosition;
 layout( location = 1 ) in vec3 vao_controlPointDirection;
-layout( location = 2 ) in unsigned short vao_ssType;
+layout( location = 2 ) in uint vao_ssType;
 layout( location = 3 ) in vec3 vao_color;
-layout( location = 4 ) in unsigned short vao_controlPointSelection;
+layout( location = 4 ) in uint vao_controlPointSelection;
 
 out VsOut
 {
 	vec4				position;
 	vec3				direction;
 	flat vec3			color;
-	flat unsigned short ssType;
-	flat unsigned short selection;
+	flat uint ssType;
+	flat uint selection;
 }
 vsOut;
